pyproject.toml
[tool.poetry]
name = "feml_heat"
version = "0.2.0"
description = ""
authors = ["horiem <yellowshippo@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.7"

matplotlib = "^3.1"
numpy = "==1.19.2"
pysiml = "==0.2.4"
sqlalchemy = "==1.3.13"
PyQt5 = {version = "^5.14.0", optional = true}
mysqlclient = {version = "^1.4.2", optional = true}
SQLAlchemy = {version = "^1.3.4", optional = true}
toml = "^0.10.2"
torch = "==1.6.0"

[tool.poetry.extras]
PyQt5 = ["PyQt5"]
mysqlclient = ["mysqlclient"]
SQLAlchemy = ["SQLAlchemy"]

[tool.poetry.dev-dependencies]
pytest = "^5.0.1"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"

[[tool.poetry.source]]
name = "ricos"
url = "https://pypi.ritc.jp/simple"
default = true
